FE/Qt: bugref:3409: Runtime UI: Fixing parfait warnings in frame-buffer code, related to implicit type conversion.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@107411 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Frontends/VirtualBox/src/runtime/UIMachineView.cpp
@@ -985,8 +985,8 @@ void UIMachineView::sltHandleActionTriggerViewScreenToggle(int iScreen, bool fEn
          * Do not send a hint if nothing has changed to prevent the guest being notified about its own changes. */
         if (   !isFullscreenOrSeamless()
             && uimachine()->isGuestSupportsGraphics()
-            && (   frameBuffer()->width() != uWidth
-                || frameBuffer()->height() != uHeight
+            && (   (ulong)frameBuffer()->width() != uWidth
+                || (ulong)frameBuffer()->height() != uHeight
                 || uimachine()->isScreenVisible(screenId()) != uimachine()->isScreenVisibleHostDesires(screenId())))
             setStoredGuestScreenSizeHint(QSize(uWidth, uHeight));
